app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# LLM + agent orchestration
openai==1.51.0
langgraph==0.2.34
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-community==0.3.1
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0

# Embeddings (local fallback)
sentence-transformers==3.1.1

# Vector store
chromadb==0.5.15
langchain-chroma==0.1.4

# Data / eval
pypdf==5.0.1
rank-bm25==0.2.2
scikit-learn==1.5.2
